pyproject.toml
[tool.poetry]
name = "dpilom310"
version = "0.1.0"
description = ""
authors = ["kifusss <kifus_06@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
setuptools = "^70.1.1"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
